feat: Add management command to clean up products
This commit introduces a new management command, delete_products, to clean the database.

The script deletes product entries with a NULL price, but only for products that have at least one other entry with a price. This is useful for removing redundant data while preserving products for which a price has not yet been found.